gmshlayerbuilder: reorganized module
