+++ scripts/gmshlayerbuilder/__main__.py
@@ -1,13 +1,15 @@
 from argparse import ArgumentParser
 
+# ensure _gmsh2meshfem is in path.
+# There may be a better way to go about this.
 try:
-    import _gmshlayerbuilder  # noqa: F401
+    import _gmsh2meshfem  # noqa: F401
 except ImportError:
     import sys
     import os
 
     sys.path.append(os.path.dirname(__file__))
-    import _gmshlayerbuilder  # noqa: F401
+    import _gmsh2meshfem  # noqa: F401
 
 
 def get_parser():
@@ -38,18 +40,19 @@ def get_parser():
 
 
 def run2D():
-    import _gmshlayerbuilder.dim2
+    import _gmsh2meshfem.dim2
+    import _gmsh2meshfem.topo_import
 
     args = get_parser().parse_args()
-    builder = _gmshlayerbuilder.dim2.layer_builder.topo_reader.builder_from_topo_file(
+    builder = _gmsh2meshfem.topo_import.builder_from_topo_file(
         args.topo_file
     )
 
     model = builder.create_model()
     if args.should_plot:
-        _gmshlayerbuilder.dim2.layer_builder.plotter.plot_model(model)
+        model.plot()
 
-    _gmshlayerbuilder.dim2.layer_builder.exporter.Exporter2D(
+    _gmsh2meshfem.dim2.Exporter(
         model, args.output_folder, nonconforming_adjacencies_file="nc_adjacencies"
     ).export_mesh()
