+++ scripts/gmshlayerbuilder/_gmsh2meshfem/dim2/__init__.py
@@ -1,0 +1,4 @@
+from .exporter import Exporter
+from . import model
+
+__all__ = ["Exporter", "model"]
