--- scripts/gmshlayerbuilder/_gmshlayerbuilder/dim2/__init__.py
@@ -1,6 +1,0 @@
-from . import layer_builder
-from .exporter import Exporter2D
-from .topo_reader import builder_from_topo_file
-from .plotter import plot_model
-
-__all__ = ["layer_builder", "Exporter2D", "builder_from_topo_file", "plot_model"]
